Move services to their module
